feat(Header): header 추출 로직 hook으로 변경
